# openapi.yaml
openapi: 3.0.3
info:
  title: CivicOS API
  version: 1.0.0
servers:
  - url: https://civicos.onrender.com
paths:
...
  /api/identity/submit:
    post:
      summary: Submit identity verification
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                termsAgreed: { type: boolean }
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized }
  /api/admin/identity-verifications:
    get:
      summary: List identity verifications (admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
...
  /api/voting/vote:
    post:
      summary: Cast a vote
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                billId: { type: integer }
                vote: { type: string, enum: [yes, no, abstain] }
      responses:
        '201': { description: Created }
        '409': { description: Already voted }
        '404': { description: Bill not found }
...
  /api/moderation/comments/recent:
    get:
      summary: List recent comments for moderation
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /api/moderation/comments/{id}:
    delete:
      summary: Remove a comment
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /api/moderation/posts/recent:
    get:
      summary: List recent posts for moderation
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /api/moderation/posts/{id}:
    delete:
      summary: Remove a post
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
        '403': { description: Forbidden }
  /api/moderation/summary:
    get:
      summary: Moderation summary counts
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT